rename: com.music.logger -> com.music.logger.service로 변경
